Introducing Spring to properly manage the lifecycle of our objects.
We want context to come to life at the beginning of each scenario and
die at the end of each scenario.

Our current tact was to hack this up, requiring the Feature file author
to be aware of this detail by demanding that they include a specific
step that rests the underlying context.

What would be better is if this behavior was automatic: that, somehow,
Step Definitions and their required context objects would be constructed
at the beginning of the scenario and disposed at the end.

That is exactly what Cucumber does with Spring.

1. When Cucumber first starts up, it instantiates a Spring Application
   Context.
2. It specifies that the main configuration lives in a file named
   "cucumber.xml" that must live at the root of the classpath.  The
   intention is for the programmer to register their custom Spring
   Beans (e.g. an object holding testing context or some other 
   resources) there.
3. Cucumber also configures the Application Context with a custom
   Scope named "cucumber-glue".  Spring Beans marked with this scope
   will be created at the beginning of a scenario and disposed at the
   end of the scenario.
4. It then registers all Step Definition classes as Spring Beans inside
   that Application Context, each in "cucumber-glue" scope.