Refactor: split concerns of runIteration()
runIteration() as written has too many concerns (over-coupling):
  1. configure and execute the next iteration,
  2. modify the stimulator's internal state with the results of the run
  3. output the results to the desire output stream.

This refactoring partially addresses that design by splitting these
concerns into separate method calls.  As a result:
- runNextIteration() does not contain any side-effects (i.e. it does
not touch the state of the stimulator object, itself).
- updateSimulatorState() does cause side-effects.  It's name indicates
that it does.  Also, this is the only method in this sequence that does
this (good cohesion).

Likely, the long-term state for this code is that these two while loops
will be collapsed into one (which incorporates the if's condition).