Major Refactor: Reorganizing our testing system
We now have enough code in our testing system that we can begin to
define the top-level partitions of the Testing System.

Summary of changes:
- Cucumber features are in two groups: Functional and GUI.  The former
houses tests that target the core functionality of the engine.  The
later verifies that the GUI operates properly.
- All test code is organized under the base package:
com.bigvisible.kanbansimulatortester.  In part, this is to convey the
design notion that our testing system is a first-class product, itself.
- JUnit tests, Cucumber Step Definitions and supporting code are
organized into product modules:
  - cli -- code that directly supports the Command Line Interface
  - common -- code reused across modules
  - core -- code that supports verifying simulator engine functionality
  - gui -- code that supports verifying GUI operation
  - suites -- test suites that aggregate different flavors of tests.