(fix) GUI tests properly handle multiple windows and tabbed panes.
+ There must be one and only one Robot for a given GUI test.  The
default constructor for FrameFixture creates a Robot; if you call that
constructor twice, the test thread will block waiting to acquire the
screen lock (a semaphore designed to prevent multiple threads from
trying to control mouse, keyboard and display.  The fix is to use one of
the other FrameFixture constructors that take an existing Robot as a
parameter.
+ By default, FEST component lookup can only find visible components. 
Tests were failing when we attempted to access the JTextArea for raw
output.  This started happening when we placed this component in a Tab
Pane and it was not the currently selected tab.

With this change, all tests are now running green.